requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
celery==5.3.1
certifi==2023.7.22
cfgv==3.4.0
charset-normalizer==3.2.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap4==2022.1
crispy-bootstrap5==0.7
decorator==5.1.1
distlib==0.3.7
Django==4.2.3
django-crispy-forms==2.0
django-debug-toolbar==4.2.0
django-environ==0.10.0
filelock==3.12.4
identify==2.5.29
idna==3.4
kombu==5.3.1
nodeenv==1.8.0
Pillow==10.0.0
platformdirs==3.10.0
pre-commit==3.4.0
prompt-toolkit==3.0.39
python-dateutil==2.8.2
PyYAML==6.0.1
requests==2.31.0
requests-mock==1.11.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
validators==0.20.0
vine==5.0.0
virtualenv==20.24.5
wcwidth==0.2.6
